pyproject.toml
[tool.poetry]
name = "finance-bot"
version = "0.1.0"
description = ""
authors = ["Джозеф <joy.life86@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
telebot = "^0.0.5"
python-dotenv = "^1.0.1"
bson = "^0.5.10"
pymongo = "^4.7.2"
decode = "^2.11.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
